move mcg_obj, created_pods, awscli_pod fixtures to tests/conftest.py
