+++ tests/manage/mcg/conftest.py
@@ -7,29 +7,11 @@ import pytest
 from botocore.exceptions import ClientError
 
 from ocs_ci.ocs import constants
-from ocs_ci.ocs.resources import mcg
 from ocs_ci.ocs.resources.mcg_bucket import S3Bucket, OCBucket, CLIBucket
 from tests import helpers
 from tests.helpers import craft_s3_command, create_unique_resource_name
 
 logger = logging.getLogger(__name__)
-
-
-@pytest.fixture()
-def mcg_obj(request):
-    """
-    Returns an MCG resource that's connected to the S3 endpoint
-    Returns:
-        MCG: An MCG resource
-
-    """
-    mcg_obj = mcg.MCG()
-
-    def finalizer():
-        mcg_obj.cred_req_obj.delete()
-    request.addfinalizer(finalizer)
-
-    return mcg_obj
 
 
 @pytest.fixture()
@@ -118,44 +100,6 @@ def bucket_factory(request, mcg_obj):
     request.addfinalizer(bucket_cleanup)
 
     return _create_buckets
-
-
-@pytest.fixture()
-def created_pods(request):
-    """
-    Deletes all pods that were created as part of the test
-
-    Returns:
-        list: An empty list of pods
-
-    """
-    created_pods_objects = []
-
-    def pod_cleanup():
-        for pod in created_pods_objects:
-            logger.info(f'Deleting pod {pod.name}')
-            pod.delete()
-    request.addfinalizer(pod_cleanup)
-    return created_pods_objects
-
-
-@pytest.fixture()
-def awscli_pod(mcg_obj, created_pods):
-    """
-    Creates a new AWSCLI pod for relaying commands
-
-    Args:
-        created_pods (Fixture/list): A fixture used to keep track of created pods
-        and clean them up in the teardown
-
-    Returns:
-        pod: A pod running the AWS CLI
-    """
-    awscli_pod_obj = helpers.create_pod(namespace=mcg_obj.namespace,
-                                        pod_dict_path=constants.AWSCLI_POD_YAML)
-    helpers.wait_for_resource_state(awscli_pod_obj, constants.STATUS_RUNNING)
-    created_pods.append(awscli_pod_obj)
-    return awscli_pod_obj
 
 
 @pytest.fixture()
